Conversion correction

+++ Phasma/Code/Shader/Shader.cpp
@@ -190,9 +190,9 @@ namespace pe
 
     std::wstring ConvertUtf8ToWide(const std::string &str)
     {
-        int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), NULL, 0);
+        int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), static_cast<int>(str.length()), NULL, 0);
         std::wstring wstr(count, 0);
-        MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), &wstr[0], count);
+        MultiByteToWideChar(CP_UTF8, 0, str.c_str(), static_cast<int>(str.length()), &wstr[0], count);
         return wstr;
     }
